Update postcard Api (#804)

+++ lib/model/trip.dart
@@ -4,7 +4,7 @@ import 'package:json_annotation/json_annotation.dart';
 class Trip {
   final String from;
   final String to;
-  final int? distance;
+  final double? distance;
 
   static Trip sendingTrip(String from) {
     return Trip(from: from, to: "Unknow", distance: null);
